+++ glean-core/android/src/main/java/mozilla/telemetry/glean/private/CustomDistributionMetricType.kt
@@ -48,7 +48,5 @@ class CustomDistributionMetricType(
     fun testGetValue(pingName: String? = null) = inner.testGetValue(pingName)
 
     @VisibleForTesting(otherwise = VisibleForTesting.NONE)
-    @JvmOverloads
-    fun testGetNumRecordedErrors(error: ErrorType, pingName: String? = null) =
-        inner.testGetNumRecordedErrors(error, pingName)
+    fun testGetNumRecordedErrors(error: ErrorType) = inner.testGetNumRecordedErrors(error)
 }