+++ glean-core/android/src/main/java/mozilla/telemetry/glean/private/MemoryDistributionMetricType.kt
@@ -34,7 +34,5 @@ class MemoryDistributionMetricType(meta: CommonMetricData, memoryUnit: MemoryUni
     fun testGetValue(pingName: String? = null) = inner.testGetValue(pingName)
 
     @VisibleForTesting(otherwise = VisibleForTesting.NONE)
-    @JvmOverloads
-    fun testGetNumRecordedErrors(error: ErrorType, pingName: String? = null) =
-        inner.testGetNumRecordedErrors(error, pingName)
+    fun testGetNumRecordedErrors(error: ErrorType) = inner.testGetNumRecordedErrors(error)
 }