+++ glean-core/python/glean/metrics/timing_distribution.py
@@ -111,9 +111,7 @@ class TimingDistributionMetricType:
         """
         return self._inner.test_get_value(ping_name)
 
-    def test_get_num_recorded_errors(
-        self, error_type: ErrorType, ping_name: Optional[str] = None
-    ) -> int:
+    def test_get_num_recorded_errors(self, error_type: ErrorType) -> int:
         """
         Returns the number of errors recorded for the given metric.
 
@@ -126,7 +124,7 @@ class TimingDistributionMetricType:
             num_errors (int): The number of errors recorded for the metric for
                 the given error type.
         """
-        return self._inner.test_get_num_recorded_errors(error_type, ping_name)
+        return self._inner.test_get_num_recorded_errors(error_type)
 
 
 __all__ = ["TimingDistributionMetricType"]