+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   * Remove `testHasValue` from all implementations.
     `testGetValue` always returns a null value
     (`null`, `nil`, `None` depending on the language) and does not throw an exception.
+* Kotlin
+  * Fix the Glean Gradle Plugin to work with Android Gradle Plugin v7.2.1 ([#2114](https://github.com/mozilla/glean/pull/2114))
 
 # v50.1.2 (2022-07-08)